Fixing the multiple products with duplicate names issue
We have several products sharing the "Tango" name -- moving the large scanner carrier to Boondock Hotel, the small box to Boondock Tango, and the scanner controller to Boondock Wilco.  While the large carrier had the name first, it's the least utilized product and will cause the least confusion amongst the user base.